aiLayer/AI layer creation/backend/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
pydantic==2.7.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
pandas
scikit-learn
huggingface-hub>=0.23.0
llama-cpp-python>=0.3.1
openai-whisper
setuptools-rust
